Move shared code into ./liveblocks
