+++ packages/liveblocks-react/src/liveblocks.tsx
@@ -26,18 +26,29 @@ import { useSyncExternalStoreWithSelector } from "use-sync-external-store/shim/w
 import { selectedInboxNotifications } from "./comments/lib/selected-inbox-notifications";
 import { retryError } from "./lib/retry-error";
 import { useInitial } from "./lib/use-initial";
-import { createSharedContext } from "./shared";
 import type {
   InboxNotificationsState,
   InboxNotificationsStateSuccess,
   LiveblocksContextBundle,
+  RoomInfoState,
+  RoomInfoStateSuccess,
+  SharedContextBundle,
   UnreadInboxNotificationsCountState,
   UnreadInboxNotificationsCountStateSuccess,
+  UserState,
+  UserStateSuccess,
 } from "./types";
 
 type OpaqueClient = Client<BaseUserMeta>;
 
 const ClientContext = createContext<OpaqueClient | null>(null);
+
+const missingUserError = new Error(
+  "resolveUsers didn't return anything for this user ID."
+);
+const missingRoomInfoError = new Error(
+  "resolveRoomsInfo didn't return anything for this room ID."
+);
 
 const _extras = new WeakMap<
   OpaqueClient,
@@ -575,6 +586,161 @@ function useCurrentUserId_withClient(client: OpaqueClient) {
     currentUserIdStore.get,
     currentUserIdStore.get
   );
+}
+
+function useUser_withClient<U extends BaseUserMeta>(
+  client: Client<U>,
+  userId: string
+): UserState<U["info"]> {
+  const usersStore = client[kInternal].usersStore;
+
+  const getUserState = useCallback(
+    () => usersStore.getState(userId),
+    [usersStore, userId]
+  );
+
+  useEffect(() => {
+    void usersStore.get(userId);
+  }, [usersStore, userId]);
+
+  const state = useSyncExternalStore(
+    usersStore.subscribe,
+    getUserState,
+    getUserState
+  );
+
+  return state
+    ? ({
+        isLoading: state.isLoading,
+        user: state.data,
+        // Return an error if `undefined` was returned by `resolveUsers` for this user ID
+        error:
+          !state.isLoading && !state.data && !state.error
+            ? missingUserError
+            : state.error,
+      } as UserState<U["info"]>)
+    : { isLoading: true };
+}
+
+function useUserSuspense_withClient<U extends BaseUserMeta>(
+  client: Client<U>,
+  userId: string
+) {
+  const usersStore = client[kInternal].usersStore;
+
+  const getUserState = useCallback(
+    () => usersStore.getState(userId),
+    [usersStore, userId]
+  );
+  const userState = getUserState();
+
+  if (!userState || userState.isLoading) {
+    throw usersStore.get(userId);
+  }
+
+  if (userState.error) {
+    throw userState.error;
+  }
+
+  // Throw an error if `undefined` was returned by `resolveUsers` for this user ID
+  if (!userState.data) {
+    throw missingUserError;
+  }
+
+  const state = useSyncExternalStore(
+    usersStore.subscribe,
+    getUserState,
+    getUserState
+  );
+
+  return {
+    isLoading: false,
+    user: state?.data,
+    error: state?.error,
+  } as UserStateSuccess<U["info"]>;
+}
+
+function useRoomInfo_withClient(client: Client, roomId: string): RoomInfoState {
+  const roomsInfoStore = client[kInternal].roomsInfoStore;
+
+  const getRoomInfoState = useCallback(
+    () => roomsInfoStore.getState(roomId),
+    [roomsInfoStore, roomId]
+  );
+
+  useEffect(() => {
+    void roomsInfoStore.get(roomId);
+  }, [roomsInfoStore, roomId]);
+
+  const state = useSyncExternalStore(
+    roomsInfoStore.subscribe,
+    getRoomInfoState,
+    getRoomInfoState
+  );
+
+  return state
+    ? ({
+        isLoading: state.isLoading,
+        info: state.data,
+        // Return an error if `undefined` was returned by `resolveRoomsInfo` for this room ID
+        error:
+          !state.isLoading && !state.data && !state.error
+            ? missingRoomInfoError
+            : state.error,
+      } as RoomInfoState)
+    : { isLoading: true };
+}
+
+function useRoomInfoSuspense_withClient(client: Client, roomId: string) {
+  const roomsInfoStore = client[kInternal].roomsInfoStore;
+
+  const getRoomInfoState = useCallback(
+    () => roomsInfoStore.getState(roomId),
+    [roomsInfoStore, roomId]
+  );
+  const roomInfoState = getRoomInfoState();
+
+  if (!roomInfoState || roomInfoState.isLoading) {
+    throw roomsInfoStore.get(roomId);
+  }
+
+  if (roomInfoState.error) {
+    throw roomInfoState.error;
+  }
+
+  // Throw an error if `undefined` was returned by `resolveRoomsInfo` for this room ID
+  if (!roomInfoState.data) {
+    throw missingRoomInfoError;
+  }
+
+  const state = useSyncExternalStore(
+    roomsInfoStore.subscribe,
+    getRoomInfoState,
+    getRoomInfoState
+  );
+
+  return {
+    isLoading: false,
+    info: state?.data,
+    error: state?.error,
+  } as RoomInfoStateSuccess;
+}
+
+/** @internal */
+export function createSharedContext<U extends BaseUserMeta>(
+  client: Client<U>
+): SharedContextBundle<U> {
+  return {
+    classic: {
+      useUser: (userId: string) => useUser_withClient(client, userId),
+      useRoomInfo: (roomId: string) => useRoomInfo_withClient(client, roomId),
+    },
+    suspense: {
+      useUser: (userId: string) => useUserSuspense_withClient(client, userId),
+      useRoomInfo: (roomId: string) =>
+        useRoomInfoSuspense_withClient(client, roomId),
+    },
+  };
 }
 
 /**
